Stuco page, slight cleanup, and start work on zapper
a bunch of stuff but mainly committing to save current zapper before
refactoring and reorganizing. Idea is for adding styles to elements or
hiding elements. Will probably be a stylesheet or list of selectors in
localstorage that can be changed or added in a gui and maybe in the
mygann+ button thing, a new button next to mygann+ settings button, or a
dropdown in the settings thing (settings, about, zapper). Shouldn't be
too hard to do just annoying sometimes dealing with blackbauds terrible
terrible css and js. its so ugly fr.

oh also stuco wanted me to add their button so I'll conform. its
appended last and toggleable in settings tho and zapper was partly
created so that stuff like it could be easily removed lol.

we be trollin, rollin, strollin, and
AHSHDASJHSAAHAHHAHAHAHHAHSDHSAHDSJAK.

cya wwwaaaaaaaaaa